cth_styledout: Common Test hook to print concise colored summary
It uses ANSI escape sequences to move the cursor and set colors.
Therefore, it might not be compatible with all terminals.

To use it:
    ct_run ... -ct_hooks cth_styledout ..

It is compatible with Erlang 19.2+ only because Common Test in previous
versions sent messages to `user` directly so we can't intercept them.
That said, it's possible to always enable this hook: it checks if it can
intercept messages and if not, the hook is a no-op.

[#138917395]